redo say error checking

+++ commands/say.ts
@@ -1,9 +1,11 @@
 import { Message } from 'discord.js'
+
 export default (message: Message, args: string[]) => {
-  if (message.guild) {
+  if (message.guild && message.guild.me.hasPermission('MANAGE_MESSAGES')) {
     message.channel.send(args.join(' '))
     message.delete()
-  } else message.channel.send('Only works in servers!')
+  } else if (message.guild) { message.channel.send('I don\'t have permission to delete messages! Give me Manage Messages!') }
+  else message.channel.send('This is a DM, I can\'t delete your message.')
 }
 
 module.exports.nodm = true
